requirements.txt
openai
python-dotenv
inquirer
subprocess